fix(onboarding): keep skip regression seam refresh-safe
